requirements.txt
eth-account
web3
python-dotenv
fastapi
uvicorn
pymongo
motor
jinja2
pydantic
python-jose[cryptography]
python-multipart
authlib
requests
pymongo[srv]
